Move chartToExplorerQueryStr to Util.ts
In order to avoid including all of bakeCovidExplorer.tsx in the client bundle.
